expense_sharing/requirements.txt
django
djangorestframework
celery